Split gradtape and keras tests
